services/ai-doc-processor/app/orchestrator/requirements.txt
boto3
strands-agents
strands-agents-tools
aws-lambda-powertools[all]
# Dependencies for AWS Textract and AWS Bedrock are included in boto3
